Remove redundant code

+++ src/main/java/configs/SpreadsheetConfigs.java
@@ -2,8 +2,8 @@ package configs;
 
 public class SpreadsheetConfigs {
 
-    public static final int DEFAULT_ROWS_NUMBER = 2;
-    public static final int DEFAULT_COLUMNS_NUMBER = 2;
+    public static final long DEFAULT_ROWS_NUMBER = 2;
+    public static final long DEFAULT_COLUMNS_NUMBER = 2;
 
     public static final String DEFAULT_SHEET_NAME = "Sheet";
